Simplified.
